polls-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-Bcrypt
PyJWT
python-dotenv
